// src/pages/product/pro.d.ts
export interface IPro {
  banners: Array<string>;
  proid: string;
  category: string;
  brand: string;
  proname: string;
  originprice: number;
  sales: number;
  stock: number;
  desc: string;
  issale: number;
  isrecommend: number;
  discount: number;
  isseckill: number;
  img1: string;
  img2: string;
  img3: string;
  img4: string;
}